pyproject.toml
[project]
name = "terrarium"
version = "0.1.0"
description = "decentralized multi-agent LLM-based framework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "fastmcp>=2.12.5",
    "matplotlib==3.10.7",
    "mcp>=1.18.0",
    "networkx>=3.3",
    "numpy>=1.26,<2.3",
    "Pillow>=12.0.0",
    "protobuf>=6.33.0",
    "python-dotenv>=1.1.1",
    "PyYAML>=6.0.3",
    "Requests>=2.32.5",
    "torch>=2.6.0",
    "tqdm>=4.66.4",
    "openai",
    "vllm==0.12.0"
]

[project.optional-dependencies]
# Microsoft Graph API integration for real Outlook/Teams calendar access
# Install with: pip install -e ".[graph]"
graph = [
    "msal>=1.31.0",        # Microsoft Authentication Library (OAuth2)
    "pytz>=2024.1",        # Timezone support
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
